% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{list_mmex_table}
\alias{list_mmex_table}
\title{List all tables in mmex db.}
\usage{
list_mmex_table(db_path)
}
\arguments{
\item{db_path}{}
}
\value{
character vector
}
\description{
List all tables in mmex db.
}
\examples{
test_db <- system.file("extdata", "mmex_test_database_2014-01-26.mmb", package = "mmex")
list_mmex_table(test_db)
}
