% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{read_mmex_db}
\alias{read_mmex_db}
\title{Read selected tables from MMEX database and combine it as a tibble.}
\usage{
read_mmex_db(db_loc)
}
\arguments{
\item{db_loc}{path to MMEX database}
}
\value{
tibble
}
\description{
Note that this does not join all tables in the db.
It only joins
\itemize{
\item checkingaccount
\item accountlist
\item category
\item subcategory
\item payee
}
}
\examples{
test_db <- system.file("extdata", "mmex_test_database_2014-01-26.mmb", package = "mmex")
df <- read_mmex_db(test_db)
}
