% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.r
\name{read_mmex_table}
\alias{read_mmex_table}
\title{Get a specific table from MMEX db using \code{query_db()}}
\usage{
read_mmex_table(tbl_name, db_path)
}
\arguments{
\item{tbl_name}{table within mmex db}

\item{db_path}{path to mmex db file}
}
\value{
tibble
}
\description{
To find the list of tables, use \code{list_mmex_table(db_path)}
}
\examples{
test_db <- system.file("extdata", "mmex_test_database_2014-01-26.mmb", package = "mmex")
df_account <- read_mmex_table('accountlist_v1', test_db)
}
